Fix broken pre-loader tests when using ICache, due to file namings
