+++ hazelcast-client/src/test/java/com/hazelcast/client/map/impl/nearcache/ClientMapNearCachePreloaderTest.java
@@ -6,7 +6,6 @@ import com.hazelcast.client.test.TestHazelcastFactory;
 import com.hazelcast.config.InMemoryFormat;
 import com.hazelcast.core.HazelcastInstance;
 import com.hazelcast.core.IMap;
-import com.hazelcast.internal.adapter.DataStructureAdapter;
 import com.hazelcast.internal.adapter.IMapDataStructureAdapter;
 import com.hazelcast.internal.nearcache.AbstractNearCachePreloaderTest;
 import com.hazelcast.internal.nearcache.NearCache;
@@ -17,19 +16,30 @@ import com.hazelcast.test.HazelcastParametersRunnerFactory;
 import com.hazelcast.test.annotation.QuickTest;
 import org.junit.After;
 import org.junit.Before;
+import org.junit.Test;
 import org.junit.experimental.categories.Category;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
 import org.junit.runners.Parameterized.Parameter;
 import org.junit.runners.Parameterized.Parameters;
 
+import java.io.File;
 import java.util.Arrays;
 import java.util.Collection;
+import java.util.concurrent.CountDownLatch;
+import java.util.concurrent.Executors;
+import java.util.concurrent.ThreadPoolExecutor;
+
+import static com.hazelcast.config.InMemoryFormat.BINARY;
+import static org.junit.Assume.assumeTrue;
 
 @RunWith(Parameterized.class)
 @Parameterized.UseParametersRunnerFactory(HazelcastParametersRunnerFactory.class)
 @Category(QuickTest.class)
 public class ClientMapNearCachePreloaderTest extends AbstractNearCachePreloaderTest<Data, String> {
+
+    private static final File DEFAULT_STORE_FILE = new File("nearCache-defaultNearCache.store").getAbsoluteFile();
+    private static final File DEFAULT_STORE_LOCK_FILE = new File(DEFAULT_STORE_FILE.getName() + ".lock").getAbsoluteFile();
 
     @Parameter
     public InMemoryFormat inMemoryFormat;
@@ -60,9 +70,55 @@ public class ClientMapNearCachePreloaderTest extends AbstractNearCachePreloaderT
         hazelcastFactory.shutdownAll();
     }
 
+    @Test
+    public void testPreloadNearCacheLock_withSharedMapConfig_concurrently()
+            throws InterruptedException {
+
+        // Ignore other memory formats, this test is not affected by that option.
+        assumeTrue(BINARY.equals(nearCacheConfig.getInMemoryFormat()));
+
+        nearCacheConfig.getPreloaderConfig().setDirectory("");
+
+        final NearCacheTestContext ctx = createContext(true);
+
+        int nThreads = 10;
+        ThreadPoolExecutor pool = (ThreadPoolExecutor) Executors.newFixedThreadPool(nThreads);
+        final CountDownLatch startLatch = new CountDownLatch(nThreads);
+        final CountDownLatch finishLatch = new CountDownLatch(nThreads);
+
+        for (int i = 0; i < nThreads; i++) {
+            pool.execute(new Runnable() {
+                @Override
+                public void run() {
+                    startLatch.countDown();
+                    try {
+                        startLatch.await();
+                    } catch (InterruptedException e) {
+                        Thread.currentThread().interrupt();
+                    }
+
+                    IMap<String, String> store = ctx.nearCacheInstance.getMap(DEFAULT_NEAR_CACHE_NAME + Thread.currentThread());
+                    for (int i = 0; i < 100; i++) {
+                        store.put("Test_" + Thread.currentThread() + "_" + i, "Value_" + Thread.currentThread() + "_" + i);
+                    }
+
+                    finishLatch.countDown();
+                }
+            });
+        }
+
+        finishLatch.await();
+        pool.shutdownNow();
+    }
+
     @Override
-    protected <K, V> DataStructureAdapter<K, V> createNewClientStore(NearCacheTestContext context, String name) {
-        return new IMapDataStructureAdapter<K, V>((IMap<K, V>) context.nearCacheInstance.getMap(name));
+    protected File getDefaultStoreFile() {
+        return DEFAULT_STORE_FILE;
+    }
+
+    @Override
+    protected File getDefaultStoreLockFile() {
+        return DEFAULT_STORE_LOCK_FILE;
     }
 
     @Override
