+++ hazelcast-client/src/test/resources/nearcache-negative-fileformat/nearCache-_hz_defaultNearCache.store
@@ -1,0 +1,1 @@
+ê<¬Nÿÿÿÿ
